added withdraw view for case where user does not reach minimum withdrawal balance, moved matic widget to withdraw component, fixed tutorial, extended timeframe for iframe task

+++ app/src/routes/user-tasks/screen/TasksConsoleItem.js
@@ -54,7 +54,7 @@ const TasksConsoleItem = ({
           </div>
         </div>
       <div style={{display: 'flex', alignItems: 'flex-end'}}>
-            <StyledGeneralParagraph paragraphColor='#00E15D' paragraphFontWeight='900'>{`${(parseInt(rewardPerClick, 10)).toFixed(2).toString()} USD`}</StyledGeneralParagraph>
+            <StyledGeneralParagraph paragraphColor='#00E15D' paragraphFontWeight='900'>{`${(parseFloat(rewardPerClick, 10)).toFixed(2).toString()} USD`}</StyledGeneralParagraph>
       </div>
       </StyledUserTaskSummaryLayout>
     </div>
